added new_mesh_from-exofile to mesh.cxx with some safety measures
